Removed enddraw with doubleclick + cleaned up the code & structure
