+++ src/index.html
@@ -21,7 +21,7 @@
 
 <div class="locateButton" onclick="setGeoLocation()"></div>
 
-<div class="measureButton" onclick="toggleMeasuring()">Klikkaa aloittaaksesi mittaus</div>
+<div class="measureButton" onclick="toggleMeasuring()">Mittaustyökalu</div>
 <!--
   <div id="search" style="position: absolute; top: .75em; right: 20em; padding: 5px; background-color: rgba(255,255,255,0.5);">
     <input type="text" placeholder="katunimihaku" style="width: 200px">
