+++ src/styles/layout.scss
@@ -27,12 +27,17 @@ html, body {
   position: absolute;
   right: 0.5em;
   bottom: auto;
-  top: 8em;
+  top: 9em;
   left: auto;
   z-index: 2;
-  width: 85px;
-  height: 60px;
+  padding: 10px 0 0 5px;
+  width: 110px;
+  height: 30px;
   background: white;
+  cursor: pointer;
+  box-shadow: 5px 5px 5px 0px rgba(0,0,0,0.75);
+  border-radius: 6px;
+  border: 1px solid #000000;
   //background-image: url("~assets/img/locate.png");
 }
